requirements.txt
# SignalStack — Python dependencies
# Pin major versions for reproducibility; patch versions float for security patches.
# Install: pip install -r requirements.txt

# ── Async / networking ────────────────────────────────────────────────────────
websockets>=12.0,<13.0
aiokafka>=0.11.0,<0.12.0
asyncpg>=0.29.0,<0.30.0
aiohttp>=3.9.0,<4.0.0

# ── Kafka admin ───────────────────────────────────────────────────────────────
kafka-python>=2.0.2,<3.0.0

# ── PySpark ───────────────────────────────────────────────────────────────────
pyspark>=3.5.0,<4.0.0

# ── Data / ML ─────────────────────────────────────────────────────────────────
numpy>=2.0.0,<3.0.0
pandas>=2.2.0,<3.0.0
yfinance>=0.2.0,<0.3.0       # for training data (free, no API key needed)
scikit-learn>=1.4.0,<2.0.0
lightgbm>=4.0.0,<5.0.0
torch>=2.3.0,<3.0.0          # for LSTM (CPU build; swap for cu121 on GPU)
mlflow>=2.13.0,<3.0.0

# ── API serving ───────────────────────────────────────────────────────────────
fastapi>=0.111.0,<0.200.0
uvicorn[standard]>=0.30.0,<0.40.0
pydantic>=2.7.0,<3.0.0

# ── Observability ─────────────────────────────────────────────────────────────
prometheus-client>=0.20.0,<0.30.0
structlog>=24.1.0,<25.0.0

# ── Config / utilities ────────────────────────────────────────────────────────
python-dotenv>=1.0.0,<2.0.0
python-dateutil>=2.9.0,<3.0.0
pyarrow>=16.0.0,<17.0.0       # Spark ↔ pandas zero-copy

# ── Polygon REST client ───────────────────────────────────────────────────────
polygon-api-client>=1.13.0,<2.0.0

# ── Dev / testing ─────────────────────────────────────────────────────────────
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.23.0,<0.30.0
pytest-mock>=3.14.0,<4.0.0
ruff>=0.4.0,<0.5.0
mypy>=1.10.0,<2.0.0